# theme/anidia-portlets/sample-vue-portlet/impl/rest-openapi.yaml
components:
  schemas:
    PostalCode:
      description: Postal code and name for municipalities covered by Anidia
      properties:
        postalCode:
          description: Postal Code
          type: string
        municipalityName:
          description: Municipality name
          type: string
        municipalityId:
          description: INE Municipality identifier
          type: string
        provinceId:
          description: Municipality name (2 first digits from postal code)
          type: string
      type: object
    Address:
      description: All the information about an address from Anidia CRM
      properties:
        kind:
          description: Type of the address (street, avenue...)
          type: string
        name:
          description: Name of the address
          type: string
    Estate:
      description: All the information about an estate
      properties:
        addressKind:
          description: Type of the address (street, avenue...)
          type: string
        addressName:
          description: Name of the address
          type: string
        number:
          description: Number of the gate
          type: string
        annex:
          description: Annex to the gate number (e.g BIS in 2 BIS)
          type: string
        gateId:
          description: Unique identifier of the gate
          type: string
    Property:
      description: All the information about the house in which someone lives
      properties:
        address:
          description: Full address for the property
          type: string
        propertyId:
          description: Property unique identifier
          type: string
        block:
          description: Block of the property (e.g. 2)
          type: string
        ladder:
          description: Ladder of the property (e.g. Izq)
          type: string
        floor:
          description: Floor of the property (e.g. 2º)
          type: string
        door:
          description: Door in the flor of the property (e.g. C)
          type: string
        status:
          description: Status of the property
          type: string
        contractStatus:
          description: Status of the contract for this property
          type: string
    Sample:
      description: A sample entity
      properties:
        name:
          description: The sample name.
          type: string
        id:
          description: The sample ID.
          type: integer
      type: object
info:
  description: "Sample API"
  license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "My Sample API for VUE"
  version: v1.0
openapi: 3.0.1
paths:
  "/properties/{gateId}":
    get:
      operationId: getPropertiesGatePage
      summary: Get all properties for an estate
      description: Get all the properties for the estate with the gate Id from Anidia CRM
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "gateId"
        in: "path"
        description: "Gate Id of the estate in CRM"
        required: true
        schema:
          type: "string"
      responses:
        200:
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Property"
            application/xml:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Property"
          description: ""
      tags: ["Property"]
  "/estates/{municipalityId}/{postalCode}/{addressKind}/{addressName}":
    get:
      operationId: getEstatesAddressNamePage
      summary: Get all estates for an address
      description: Get all the estates for a whole address from Anidia CRM
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "municipalityId"
        in: "path"
        description: "INE Municipality identifier of the address"
        required: true
        schema:
          type: "string"
      - name: "postalCode"
        in: "path"
        description: "Postal Code of the address"
        required: true
        schema:
          type: "string"
      - name: "addressKind"
        in: "path"
        description: "Kind of the address"
        required: true
        schema:
          type: "string"
      - name: "addressName"
        in: "path"
        description: "Name of the address"
        required: true
        schema:
          type: "string"
      responses:
        200:
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Estate"
            application/xml:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Estate"
          description: ""
      tags: ["Estate"]
  "/addresses/{municipalityId}/{postalCode}":
    get:
      operationId: getAddressesPostalCodePage
      summary: Get all addresses for a municipality and postal code
      description: Get all the addresses for a municipality and postal code from Anidia CRM
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "municipalityId"
        in: "path"
        description: "INE Municipality identifier of the addresses"
        required: true
        schema:
          type: "string"
      - name: "postalCode"
        in: "path"
        description: "Postal Code of the addresses"
        required: true
        schema:
          type: "string"
      responses:
        200:
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Address"
            application/xml:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Address"
          description: ""
      tags: ["Address"]

  "/postal-codes":
    get:
      operationId: getPostalCodesPage
      summary: Get all postal codes
      description: Get all the postal codes from the Anidia database
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PostalCode"
            application/xml:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PostalCode"
          description: ""
      tags: ["PostalCode"]

  "/samples/{sampleId}":
    get:
      operationId: getSample
      parameters:
        - in: path
          name: sampleId
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sample"
            application/xml:
              schema:
                $ref: "#/components/schemas/Sample"
          description: ""
      tags: ["Sample"]